+++ packages/playground/eslint.config.mjs
@@ -1,15 +1,18 @@
 // @ts-check
 
 import js from "@eslint/js"
+import reactHooks from "eslint-plugin-react-hooks"
+import reactRefresh from "eslint-plugin-react-refresh"
 import { defineConfig, globalIgnores } from "eslint/config"
 import globals from "globals"
 import tseslint from "typescript-eslint"
 
 export default defineConfig([
     globalIgnores(["node_modules", "dist", "build", "public"]),
+    reactHooks.configs.flat.recommended,
     {
-        files: ["**/*.{js,mjs,ts}"],
-        extends: [js.configs.recommended, tseslint.configs.recommended],
+        files: ["**/*.{js,mjs,ts,tsx}"],
+        extends: [js.configs.recommended, tseslint.configs.recommended, reactRefresh.configs.vite],
         languageOptions: {
             ecmaVersion: "latest",
             globals: globals.browser,
@@ -22,7 +25,7 @@ export default defineConfig([
             "no-empty": "off",
             "no-extra-boolean-cast": "off",
             "no-unused-vars": "off",
-            "@typescript-eslint/no-explicit-any": "off",
+            "react-refresh/only-export-components": "warn",
             "@typescript-eslint/no-unused-vars": [
                 "warn",
                 {